feat: delete and merge innovation tiers migration
